requirements.txt
Flask==2.2.5
Werkzeug==3.1.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Limiter==3.3.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.0.0
python-dateutil==2.8.2
pyotp==2.8.0
qrcode==7.4.2
Pillow==10.0.0
plotly==5.14.1
pandas==2.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
pytest==7.4.0
pytest-cov==4.1.0
